Jiahaochen-personal_site_streamlit/requirements.txt
streamlit
pillow
matplot
pandas
numpy